assertDNSRecord: Increase timeout to 10m
Increase the timeout in assertDNSRecord for polling for the DNSRecord CR
from 1 minute to 10 minutes.

The cloud provider can easily take over a minute to provision the load balancer,
and the operator cannot create the DNSRecord CR before the load balancer has
been provisioned and assigned a host name or address.  Consequently, the polling
loop could easily reach the 1-minute timeout just on account of the time that it
takes to provision the load balancer.

* test/e2e/util_gatewayapi_test.go (assertDNSRecord): Increase timeout
for the DNSRecord CR polling loop from 1m to 10m.